Streamline the README. Minor improvements
